PG_test/requirements.txt
Django==2.2.7
django-cors-headers==3.1.1
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1
psycopg2==2.8.4
pytz==2019.1
sqlparse==0.3.0
